fix: UTC time, int64 timestamps, double wallet, correct range generator
The numeric-drift half of the review fixes, kept separate from the
bit-identical batch so the fixture change is reviewable on its own.
Unit suite 100 -> 122 checks.

generateRange_int computed its step with integer division, so the range
never reached vmax: generateRange_int(3, 600, 300) produced 3..302 rather
than 3..600, meaning every F_* sweep silently searched about half the
parameter space it advertised. N == 1 also divided by zero. It now
interpolates in floating point, always includes both endpoints, and
deduplicates when the interval is narrower than N.

The calendar helpers used localtime(), which made every result depend on
the host timezone -- the same data and parameters gave different
year/month/day boundaries, and so different Calmar ratios and month
rollovers, on two machines. localtime()/gmtime() also return a pointer to
a shared static tm that the two F_* worker threads raced on. All of them
now use gmtime_r and read tm fields directly, dropping a
strftime-then-stoi round-trip that ran once or twice per candle in the
hot loops. Exchange klines are UTC, so UTC is the correct reading.

Timestamps are int64_t end to end (KLINEf, fundings, WalletTrace, the
calendar helpers, funding lookup). They were narrowed to int at nearly
every use, which overflows in January 2038. 3EMA_SRSI_ATR and
F_3EMA_SRSI_ATR also held position-open timestamps in an unsigned array
while assigning -10 as a cleared sentinel, which wrapped to a huge
positive value.

Wallet, fees, position sizes and all result metrics are double; prices
and indicator series stay float, where the memory win actually is. The
wallet is an accumulator -- thousands of fee subtractions and position
round-trips were spending float's ~7 significant digits on the very
totals the sweep ranks candidates by. A new test runs 5000 round-trips
against the closed form and requires relative error below 1e-12; the
existing fee tests tighten from 1e-2 to 1e-9.

calculate_calmar_ratio's third parameter is renamed from max_DD to
drawdown_normalizer: every caller passes ResultMetrics::ddc, a positive
number, not the negative max drawdown. Both Calmar functions now guard a
non-positive normalizer and an empty change list instead of dividing by
zero, and four unused locals are gone.

Fixture change: strategy_regression only. Trade count (166) and the
selected EMA pair are identical; gain, win rate, max DD, gain/DDC and
score move in the 7th significant digit, which is the float accumulation
error being removed. verification_regression is unchanged (indicators are
still float) and backtest_double_EMA_float still selects EMA 448/11 with
identical printed metrics across its full 175k-point sweep.

+++ 3EMA_SRSI_ATR.cpp
@@ -95,7 +95,7 @@ RUN_RESULTf PROCESS(const std::vector<KLINEf> &PAIRS, const int &ema1, const int
 
     bool LAST_ITERATION = false;
     array<float, NB_PAIRS> ATR_AT_OPEN{};
-    array<uint, NB_PAIRS> OPEN_TS{};
+    array<int64_t, NB_PAIRS> OPEN_TS{};
 
     const uint ii_begin = start_indexes[0];
 
@@ -187,7 +187,7 @@ RUN_RESULTf PROCESS(const std::vector<KLINEf> &PAIRS, const int &ema1, const int
         last_closes[ic] = PAIRS[ic].close[nb_max - 1];
     }
 
-    const float wallet_val_usdt = trade_core::calculate_spot_wallet_val_usdt(portfolio, last_closes);
+    const double wallet_val_usdt = trade_core::calculate_spot_wallet_val_usdt(portfolio, last_closes);
 
     const trade_core::ResultMetrics metrics = trade_core::calculate_result_metrics(wallet_val_usdt, USDT_amount_initial, portfolio.max_drawdown, stats);
