+++ BBTREND.cpp
@@ -153,7 +153,7 @@ RUN_RESULTf PROCESS(const vector<KLINEf> &PAIRS, const int &ema_v, const int &BB
         last_closes[ic] = PAIRS[ic].close[nb_max - 1];
     }
 
-    const float wallet_val_usdt = trade_core::calculate_spot_wallet_val_usdt(portfolio, last_closes);
+    const double wallet_val_usdt = trade_core::calculate_spot_wallet_val_usdt(portfolio, last_closes);
 
     const trade_core::ResultMetrics metrics = trade_core::calculate_result_metrics(wallet_val_usdt, USDT_amount_initial, portfolio.max_drawdown, stats);
